cleanup tests, make pkce default for server
